// src/api/openapi/auto-generated/models/DisclosureCategory.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const DisclosureCategory = {
    temperature: 'temperature',
    wind: 'wind',
    water: 'water',
    solidMass: 'solid_mass',
    unknownDefaultOpenApi: '11184809'
} as const;
export type DisclosureCategory = typeof DisclosureCategory[keyof typeof DisclosureCategory];


export function instanceOfDisclosureCategory(value: any): boolean {
    for (const key in DisclosureCategory) {
        if (Object.prototype.hasOwnProperty.call(DisclosureCategory, key)) {
            if ((DisclosureCategory as Record<string, DisclosureCategory>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function DisclosureCategoryFromJSON(json: any): DisclosureCategory {
    return DisclosureCategoryFromJSONTyped(json, false);
}

export function DisclosureCategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): DisclosureCategory {
    return json as DisclosureCategory;
}

export function DisclosureCategoryToJSON(value?: DisclosureCategory | null): any {
    return value as any;
}